requirements.txt
numpy
opencv-python
pyrender
pytorch-lightning
scikit-image
smplx==0.1.28
yacs
chumpy @ git+https://github.com/mattloper/chumpy
timm
einops
xtcocotools
pandas
hydra-core
hydra-submitit-launcher
hydra-colorlog
pyrootutils
rich
webdataset
gradio